requirements.txt
alembic==1.13.2
fastapi==0.112.2
httpx==0.27.2
psycopg2-binary==2.9.9
pydantic==2.8.2
PyJWT==2.9.0
pytest==8.3.2
python-decouple==3.8
python-multipart==0.0.9
pyyaml==6.0.2
SQLAlchemy==2.0.32
uvicorn==0.30.6
